Centralize vehicle physical status logic in one pure helper
