+++ apps/lrauv-dash2/lib/useVehicleStatus.ts
@@ -1,10 +1,8 @@
-import { useMemo } from 'react'
 import type { DeploymentEvent } from '@mbari/api-client'
 import { useTethysSubscriptionEvent } from './useWebSocketListeners'
-import { isEffectivelyCommunicating } from './effectiveVehicleComms'
-import { getPluggedInStatus } from './getPluggedInStatus'
+import { getVehiclePhysicalStatus } from './getVehiclePhysicalStatus'
 
-export type VehiclePhysicalStatus = 'pluggedIn' | 'surfaced' | 'underwater'
+export type { VehiclePhysicalStatus } from './getVehiclePhysicalStatus'
 
 export type UseVehicleStatusArgs = {
   vehicleName: string | undefined
@@ -27,35 +25,16 @@ export const useVehicleStatus = ({
     'VehiclePingResult',
     vehicleName ?? ''
   )
-  const cellPingReachable = Boolean(pingEvent?.reachable)
-
-  const isPluggedIn = getPluggedInStatus({
-    recoverEvent,
-    startEventUnix,
-  })
-
-  const isLikelySurfaced = useMemo(
-    () =>
-      isEffectivelyCommunicating({
-        pingReachable: cellPingReachable,
-        lastSatCommsTime,
-        lastCellCommsTime,
-        nowMs,
-      }),
-    [cellPingReachable, lastSatCommsTime, lastCellCommsTime, nowMs]
-  )
-
-  const physicalStatus: VehiclePhysicalStatus = isPluggedIn
-    ? 'pluggedIn'
-    : isLikelySurfaced
-    ? 'surfaced'
-    : 'underwater'
 
   return {
     pingEvent,
-    cellPingReachable,
-    isPluggedIn,
-    isLikelySurfaced,
-    physicalStatus,
+    ...getVehiclePhysicalStatus({
+      pingReachable: pingEvent?.reachable,
+      lastSatCommsTime,
+      lastCellCommsTime,
+      nowMs,
+      recoverEvent,
+      startEventUnix,
+    }),
   }
 }
